requirements.txt
deepl
groq
keyboard
pyautogui
pyperclip
win32gui
win32console
azure-cognitiveservices-speech